# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AR Studio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Folder {
  id: number;
  name: string;
  /** @nullable */
  parentId: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFolderBody {
  name: string;
  /** @nullable */
  parentId?: number | null;
}

export interface UpdateFolderBody {
  name?: string;
}

export type ProjectEnvironment =
  (typeof ProjectEnvironment)[keyof typeof ProjectEnvironment];

export const ProjectEnvironment = {
  black: "black",
  white: "white",
  "luxury-home": "luxury-home",
  "classic-luxury": "classic-luxury",
  "walls-plants": "walls-plants",
} as const;

export type ProjectType = (typeof ProjectType)[keyof typeof ProjectType];

export const ProjectType = {
  furniture: "furniture",
  object: "object",
} as const;

export interface Project {
  id: number;
  name: string;
  companyName: string;
  /** @nullable */
  thumbnail: string | null;
  /** @nullable */
  modelUrl: string | null;
  isLive: boolean;
  environment: ProjectEnvironment;
  hotspotX: number;
  hotspotY: number;
  hotspotZ: number;
  language: string;
  type: ProjectType;
  isScalable: boolean;
  enableMaterials: boolean;
  enableVariants: boolean;
  /** @nullable */
  folderId: number | null;
  publicSlug: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateProjectBodyEnvironment =
  (typeof CreateProjectBodyEnvironment)[keyof typeof CreateProjectBodyEnvironment];

export const CreateProjectBodyEnvironment = {
  black: "black",
  white: "white",
  "luxury-home": "luxury-home",
  "classic-luxury": "classic-luxury",
  "walls-plants": "walls-plants",
} as const;

export type CreateProjectBodyType =
  (typeof CreateProjectBodyType)[keyof typeof CreateProjectBodyType];

export const CreateProjectBodyType = {
  furniture: "furniture",
  object: "object",
} as const;

export interface CreateProjectBody {
  name: string;
  companyName: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  environment?: CreateProjectBodyEnvironment;
  hotspotX?: number;
  hotspotY?: number;
  hotspotZ?: number;
  language?: string;
  type?: CreateProjectBodyType;
  isScalable?: boolean;
  enableMaterials?: boolean;
  enableVariants?: boolean;
  /** @nullable */
  folderId?: number | null;
}

export type UpdateProjectBodyEnvironment =
  (typeof UpdateProjectBodyEnvironment)[keyof typeof UpdateProjectBodyEnvironment];

export const UpdateProjectBodyEnvironment = {
  black: "black",
  white: "white",
  "luxury-home": "luxury-home",
  "classic-luxury": "classic-luxury",
  "walls-plants": "walls-plants",
} as const;

export type UpdateProjectBodyType =
  (typeof UpdateProjectBodyType)[keyof typeof UpdateProjectBodyType];

export const UpdateProjectBodyType = {
  furniture: "furniture",
  object: "object",
} as const;

export interface UpdateProjectBody {
  name?: string;
  companyName?: string;
  /** @nullable */
  thumbnail?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  environment?: UpdateProjectBodyEnvironment;
  hotspotX?: number;
  hotspotY?: number;
  hotspotZ?: number;
  language?: string;
  type?: UpdateProjectBodyType;
  isScalable?: boolean;
  enableMaterials?: boolean;
  enableVariants?: boolean;
  /** @nullable */
  folderId?: number | null;
}

export type StudioProjectEnvironment =
  (typeof StudioProjectEnvironment)[keyof typeof StudioProjectEnvironment];

export const StudioProjectEnvironment = {
  black: "black",
  white: "white",
  "luxury-home": "luxury-home",
  "classic-luxury": "classic-luxury",
  "walls-plants": "walls-plants",
} as const;

export type StudioProjectType =
  (typeof StudioProjectType)[keyof typeof StudioProjectType];

export const StudioProjectType = {
  furniture: "furniture",
  object: "object",
} as const;

export interface ProjectMaterial {
  id: number;
  projectId: number;
  name: string;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  modelUrl: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface ProjectVariant {
  id: number;
  projectId: number;
  name: string;
  /** @nullable */
  thumbnailUrl: string | null;
  /** @nullable */
  modelUrl: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface StudioProject {
  id: number;
  name: string;
  companyName: string;
  /** @nullable */
  modelUrl: string | null;
  environment: StudioProjectEnvironment;
  hotspotX: number;
  hotspotY: number;
  hotspotZ: number;
  language: string;
  type: StudioProjectType;
  isScalable: boolean;
  enableMaterials: boolean;
  enableVariants: boolean;
  publicSlug: string;
  materials: ProjectMaterial[];
  variants: ProjectVariant[];
}

export interface EnvironmentCount {
  environment: string;
  count: number;
}

export interface DashboardStats {
  totalProjects: number;
  liveARs: number;
  avgLoadSpeedMs: number;
  recentProjects: Project[];
  projectsByEnvironment: EnvironmentCount[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface CreateMaterialBody {
  name: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  sortOrder?: number;
}

export interface UpdateMaterialBody {
  name?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  sortOrder?: number;
}

export interface CreateVariantBody {
  name: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  sortOrder?: number;
}

export interface UpdateVariantBody {
  name?: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  sortOrder?: number;
}

export type ListProjectsParams = {
  /**
   * @nullable
   */
  folderId?: number | null;
};