# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AR Studio API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: folderId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new AR project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/publish:
    post:
      operationId: publishProject
      tags: [projects]
      summary: Publish (activate) a project AR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project published with permanent link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
  /projects/{id}/unpublish:
    post:
      operationId: unpublishProject
      tags: [projects]
      summary: Unpublish (deactivate) a project AR
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project unpublished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found

  # ── Public Studio ────────────────────────────────────────────────────────────
  /studio/{slug}:
    get:
      operationId: getStudioProject
      tags: [projects]
      summary: Get public studio data for a project (no auth required)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Studio project data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudioProject"
        "404":
          description: Not found

  # ── Stats ────────────────────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL for a 3D model file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve an uploaded object file
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object file content
        "404":
          description: Not found

  # ── Materials ──────────────────────────────────────────────────────────────
  /projects/{projectId}/materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List materials for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of materials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMaterial"

    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create a material for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMaterialBody"
      responses:
        "201":
          description: Material created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMaterial"
  /projects/{projectId}/materials/{id}:
    patch:
      operationId: updateMaterial
      tags: [materials]
      summary: Update a material
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMaterialBody"
      responses:
        "200":
          description: Updated material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMaterial"
        "404":
          description: Not found

    delete:
      operationId: deleteMaterial
      tags: [materials]
      summary: Delete a material
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Variants ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: string
        thumbnail:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        isLive:
          type: boolean
        environment:
          type: string
          enum: [black, white, luxury-home, classic-luxury, walls-plants]
        hotspotX:
          type: number
        hotspotY:
          type: number
        hotspotZ:
          type: number
        language:
          type: string
        type:
          type: string
          enum: [furniture, object]
        isScalable:
          type: boolean
        enableMaterials:
          type: boolean
        enableVariants:
          type: boolean
        folderId:
          type: ["integer", "null"]
        publicSlug:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - companyName
        - thumbnail
        - modelUrl
        - isLive
        - environment
        - hotspotX
        - hotspotY
        - hotspotZ
        - language
        - type
        - isScalable
        - enableMaterials
        - enableVariants
        - folderId
        - publicSlug
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        thumbnail:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        environment:
          type: string
          enum: [black, white, luxury-home, classic-luxury, walls-plants]
        hotspotX:
          type: number
        hotspotY:
          type: number
        hotspotZ:
          type: number
        language:
          type: string
        type:
          type: string
          enum: [furniture, object]
        isScalable:
          type: boolean
        enableMaterials:
          type: boolean
        enableVariants:
          type: boolean
        folderId:
          type: ["integer", "null"]
      required:
        - name
        - companyName
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        companyName:
          type: string
        thumbnail:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        environment:
          type: string
          enum: [black, white, luxury-home, classic-luxury, walls-plants]
        hotspotX:
          type: number
        hotspotY:
          type: number
        hotspotZ:
          type: number
        language:
          type: string
        type:
          type: string
          enum: [furniture, object]
        isScalable:
          type: boolean
        enableMaterials:
          type: boolean
        enableVariants:
          type: boolean
        folderId:
          type: ["integer", "null"]
    StudioProject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        companyName:
          type: string
        modelUrl:
          type: ["string", "null"]
        environment:
          type: string
          enum: [black, white, luxury-home, classic-luxury, walls-plants]
        hotspotX:
          type: number
        hotspotY:
          type: number
        hotspotZ:
          type: number
        language:
          type: string
        type:
          type: string
          enum: [furniture, object]
        isScalable:
          type: boolean
        enableMaterials:
          type: boolean
        enableVariants:
          type: boolean
        publicSlug:
          type: string
        materials:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMaterial"
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProjectVariant"
      required:
        - id
        - name
        - companyName
        - modelUrl
        - environment
        - hotspotX
        - hotspotY
        - hotspotZ
        - language
        - type
        - isScalable
        - enableMaterials
        - enableVariants
        - publicSlug
        - materials
        - variants
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required:
        - uploadURL
        - objectPath
    ProjectMaterial:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - thumbnailUrl
        - modelUrl
        - sortOrder
        - createdAt
    CreateMaterialBody:
      type: object
      properties:
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - name
    UpdateMaterialBody:
      type: object
      properties:
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    ProjectVariant:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        modelUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - thumbnailUrl
        - modelUrl
        - sortOrder
        - createdAt
...