# lib/api-zod/src/generated/types/createMaterialBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AR Studio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMaterialBody {
  name: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  modelUrl?: string | null;
  sortOrder?: number;
}